chore: use published blstrs and decouple deps
-- Change 1 --

The idea is to give API consumers the option to use the exported crates
without needing explicit Cargo.toml entries.

This in turn decouples dependencies and reduces conflicts.

See rationale/discussion here:
https://github.com/rust-lang/api-guidelines/discussions/176

-- Change 2 --

We use blstrs from bulletproofs, which is now using the published
blstrs.  To use the published blstrs, we had to change ::sum()
calls for Scalar to use ::fold() instead.

We can change it back to ::sum() if/when published blstrs accepts
our PR for that.

Changes:
* re-export deps that are used in our public apis
* use deps from bulletproofs (blstrs, group, merlin, rand-core)
* replace ::sum() with ::fold() for Scalar

Cargo.toml:
* use blst-bulletproofs that re-exports public api deps
* remove deps that are re-exported in blst-bulletproofs