fix: verify pubkey are unique across inputs
Issue: an sn_dbc test case was creating a Tx with
an input mlsag that had two duplicate public_keys.

It was determined this was occurring because
SpentbookNodeMock::random_decoy() returns a random set of decoys, which
in this case included the true input.

RingCtTransaction::verify() did not return any error for this
transaction and the test case passed when it should not have.

The fix is to validate that the public key is unique for every
possible input (true and decoys) across all mlsag.

Changes:
* add Error variant PublicKeyNotUniqueAcrossInputs
* verify that public keys are unique across all input mlsag