chore: rename crate to bls_ringct
Yesterday we renamed this crate from blst_ringct to sn_ringct, but later we decided it would be
independent of Safe, so we'll rename again. We lose the 'sn' prefix in favour of 'bls', which drops
the 't' from the original name, since that's a specific implementation detail of the original
repository, which our fork has now diverged from. We also have other repos/crates that use this
'bls' prefix, so it brings it in line with those.

We also replace the `sn_bulletproofs` reference to `bls_bulletproofs`, which was a crate that was
also renamed.